Update button handling and improve game actions
In this commit, we've modified the handling of buttons in the code using the Bounce2 library, which has been added to our dependencies. The code now uses this library to capture button press events more efficiently. We've also made adjustments to the different game actions in the source file, enhancing the sequence and dynamics of the keyboard and mouse commands.